+++ app/build.gradle.kts
@@ -71,6 +71,7 @@ android {
 
 dependencies {
     implementation(projects.core.android)
+    implementation(projects.core.androidx.compose)
     implementation(projects.core.androidx.fragment)
     implementation(projects.core.androidx.lifecycle)
     implementation(projects.core.androidx.navigationCompose)
@@ -92,7 +93,6 @@ dependencies {
     implementation(projects.core.statekit)
     kapt(projects.statekit.compiler)
     implementation(projects.core.translation)
-    implementation(projects.core.uiCompose)
     implementation(projects.core.uitext)
 
     implementation(libs.kotlinx.coroutines.android)
