+++ core/androidx/compose/src/main/AndroidManifest.xml
