+++ core/androidx/compose/src/main/kotlin/com/nlab/reminder/core/androidx/compose/ui/tooling/preview/Previews.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.ui.compose.tooling.preview
+package com.nlab.reminder.core.androidx.compose.ui.tooling.preview
 
 import android.content.res.Configuration.UI_MODE_NIGHT_NO
 import android.content.res.Configuration.UI_MODE_NIGHT_YES
